Move Product Statuses to constants
