+++ projects/packages/my-jetpack/_inc/components/connected-product-card/stories/broken/index.stories.jsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { HashRouter, Routes, Route } from 'react-router-dom';
 import ConnectedProductCard from '../';
-import { PRODUCT_STATUSES } from '../../product-card';
+import { PRODUCT_STATUSES } from '../../../../constants';
 import { getAllMockData, getProductSlugs } from '../../product-detail-card/stories/utils.js';
 
 // Set myJetpackRest global var.
