+++ projects/packages/my-jetpack/_inc/constants.ts
@@ -23,3 +23,17 @@ export const MyJetpackRoutes = {
 	AddLicense: '/add-license',
 	RedeemToken: '/redeem-token',
 } as const;
+
+export const PRODUCT_STATUSES = {
+	ACTIVE: 'active',
+	INACTIVE: 'inactive',
+	MODULE_DISABLED: 'module_disabled',
+	SITE_CONNECTION_ERROR: 'site_connection_error',
+	ABSENT: 'plugin_absent',
+	ABSENT_WITH_PLAN: 'plugin_absent_with_plan',
+	NEEDS_PURCHASE: 'needs_purchase',
+	NEEDS_PURCHASE_OR_FREE: 'needs_purchase_or_free',
+	NEEDS_FIRST_SITE_CONNECTION: 'needs_first_site_connection',
+	USER_CONNECTION_ERROR: 'user_connection_error',
+	CAN_UPGRADE: 'can_upgrade',
+};
