+++ projects/packages/my-jetpack/_inc/data/utils/get-product-slugs-that-require-user-connection.ts
@@ -1,4 +1,4 @@
-import { PRODUCT_STATUSES } from '../../components/product-card';
+import { PRODUCT_STATUSES } from '../../constants';
 import type { ProductCamelCase } from '../types';
 
 const getProductSlugsThatRequireUserConnection = ( products: {
